requirements.txt
# AgriSahayak Backend Requirements
# For deployment (CPU-optimized)

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# ML & AI (CPU versions for deployment)
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
httpx==0.26.0
joblib==1.3.2
python-dotenv==1.0.0
